SoundManagerMaster Commit
Made the SoundManagerMaster parent of 5 sound sources. The code will
cycle through the array of sound sources to play as the overall
manager, the number system itself is currently flawed, so it is unable
to play more than 1 sound at a time, so this will need to be worked on.
Of course SoundManager.cs may be used to reroute sounds to a
SoundManager using
SoundManager.instance.NameOfMethodHere(nameOfFileHoldingParameterHere);